requirements.txt
numpy>=1.19.0
Pillow>=8.0.0
PyOpenGL>=3.1.0
PyOpenGL-accelerate>=3.1.0
pyopengltk>=0.0.4 